+++ ofxGgmlMontagePlannerExample/src/main.cpp
@@ -1,0 +1,7 @@
+#include "ofMain.h"
+#include "ofApp.h"
+
+int main() {
+ofSetupOpenGL(1180, 700, OF_WINDOW);
+ofRunApp(new ofApp());
+}